requirements.txt
Django==4.2.7
django-ninja==0.22.2
greenlet==3.0.1
gunicorn==21.2.0
mssql-django==1.3
psycopg[binary]==3.1.16
pydantic[email]~=1.10.13
pyodbc==5.0.1
uvicorn==0.24.0.post1
# библиотеки для flake8 и mypy
mypy==1.7.1
flake8==6.1.0
flake8-docstrings==1.7.0
flake8-import-order==0.18.2
flake8-quotes==3.3.2